Crate bcrypt

Easily hash and verify passwords using bcrypt

Structs

HashParts

A bcrypt hash result before concatenating

Enums

BcryptError

All the errors we can encounter while hashing/verifying passwords

Version

BCrypt hash version https://en.wikipedia.org/wiki/Bcrypt#Versioning_history

Constants

BASE_64

DEFAULT_COST

Functions

bcrypt

hash

Generates a password hash using the cost given. The salt is generated randomly using the OS randomness

hash_with_result

Generates a password hash using the cost given. The salt is generated randomly using the OS randomness. The function returns a result structure and allows to format the hash in different versions.

hash_with_salt

Generates a password given a hash and a cost. The function returns a result structure and allows to format the hash in different versions.

non_truncating_hash

Generates a password hash using the cost given. The salt is generated randomly using the OS randomness Will return BcryptError::Truncation if password is longer than 72 bytes

non_truncating_hash_with_result

Generates a password hash using the cost given. The salt is generated randomly using the OS randomness. The function returns a result structure and allows to format the hash in different versions. Will return BcryptError::Truncation if password is longer than 72 bytes

non_truncating_hash_with_salt

Generates a password given a hash and a cost. The function returns a result structure and allows to format the hash in different versions. Will return BcryptError::Truncation if password is longer than 72 bytes

non_truncating_verify

Verify that a password is equivalent to the hash provided. Only use this if you are only using non_truncating_hash to generate the hash. It will return an error for inputs that will work if generated from other sources.

verify

Verify that a password is equivalent to the hash provided

Type Aliases

BcryptResult

Library generic result type.