Crate bcrypt

Easily hash and verify passwords using bcrypt

Structs

HashParts
A bcrypt hash result before concatenating

Enums

BcryptError
All the errors we can encounter while hashing/verifying passwords

Version
BCrypt hash version https://en.wikipedia.org/wiki/Bcrypt#Versioning_history

Constants

DEFAULT_COST

Functions

bcrypt

hash
Generates a password hash using the cost given. The salt is generated randomly using the OS randomness

hash_with_result
Generates a password hash using the cost given. The salt is generated randomly using the OS randomness. The function returns a result structure and allows to format the hash in different versions.

hash_with_salt
Generates a password given a hash and a cost. The function returns a result structure and allows to format the hash in different versions.

verify
Verify that a password is equivalent to the hash provided

Type Definitions

BcryptResult
Library generic result type.